## Setting version of Kwave in the Cmake files
This would help give the app stream metadata. As these applications like Discover use this version as metadata for a listing of the application.

This is a fix proposed for the following bug pointed to on Bugzilla: https://bugs.kde.org/show_bug.cgi?id=425034

A similar change was also proposed to Kdenline for similar reasons, which can be seen here: https://github.com/KDE/kdenlive/commit/50085b9fca6690e22145a0afbf3882746877bf68