Use the proper target names when linking
Fixes the build, which otherwise fails with
"ld: cannot find -lkwave: No such file or directory".

I suspect this broke by requiring a newer ECM version and newer cmake
policies, but it's the correct thing to do anyway.
The add_dependencies invocations look like an earlier try to fix the
symptoms of using the actual, verbatim library names instead of the
target names. They shouldn't be necessary anymore because cmake
automatically handles the dependencies now that it knows about the
targets.